# shared/sdk-ts/openapi.json
{
  "components": {
    "schemas": {
...
      "DictamenList": {
        "description": "Cadena de dictámenes de un incidente (más reciente primero).",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DictamenOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "DictamenList",
        "type": "object"
      },
      "DictamenOut": {
        "description": "Fila de ``dictamens``: ``signed_by`` NULL = preliminar; cadena vía supersedes.",
        "properties": {
          "basis": {
            "additionalProperties": true,
            "title": "Basis",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dictamen_id": {
            "format": "uuid",
            "title": "Dictamen Id",
            "type": "string"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "signed_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signed By"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "supersedes_dictamen_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supersedes Dictamen Id"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "dictamen_id",
          "tenant_id",
          "incident_id",
          "status",
          "basis",
          "signed_by",
          "supersedes_dictamen_id",
          "created_at"
        ],
        "title": "DictamenOut",
        "type": "object"
      },
      "DictamenSignIn": {
        "description": "Firma manual del inspector: estado del dictamen + notas mínimas (basis).",
        "properties": {
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "DictamenSignIn",
        "type": "object"
      },
...
      "EventDetailOut": {
        "description": "Detalle de evento: campos del evento + sus votos de quórum.",
        "properties": {
          "depth_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Depth Km"
          },
          "detected_at": {
            "format": "date-time",
            "title": "Detected At",
            "type": "string"
          },
          "epicenter_lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Epicenter Lat"
          },
          "epicenter_lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Epicenter Lon"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "magnitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Magnitude"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          },
          "quorum_votes": {
            "items": {
              "$ref": "#/components/schemas/QuorumVoteOut"
            },
            "title": "Quorum Votes",
            "type": "array"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "source",
          "magnitude",
          "epicenter_lon",
          "epicenter_lat",
          "depth_km",
          "detected_at",
          "meta",
          "quorum_votes"
        ],
        "title": "EventDetailOut",
        "type": "object"
      },
      "EventPage": {
        "description": "Página keyset de eventos sísmicos.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SeismicEventOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "EventPage",
        "type": "object"
      },
      "EvidenceList": {
        "description": "Lista de evidencias de un incidente (ya filtrada por RLS).",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EvidenceObject"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "EvidenceList",
        "type": "object"
      },
      "EvidenceObject": {
        "description": "Puntero a un objeto de evidencia en S3 (metadatos; el binario vive en S3).",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evidence_id": {
            "format": "uuid",
            "title": "Evidence Id",
            "type": "string"
          },
          "incident_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Id"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "s3_key": {
            "title": "S3 Key",
            "type": "string"
          },
          "sensor_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensor Id"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "ts_from": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ts From"
          },
          "ts_to": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ts To"
          }
        },
        "required": [
          "evidence_id",
          "kind",
          "s3_key",
          "created_at"
        ],
        "title": "EvidenceObject",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PresignedDownload": {
        "description": "URL GET presignada de vida corta para descargar un objeto de evidencia.",
        "properties": {
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "expires_in"
        ],
        "title": "PresignedDownload",
        "type": "object"
      },
      "QuorumVoteOut": {
        "description": "Voto de ``quorum_votes``: arribo por sensor/estación con ``delta_s``.",
        "properties": {
          "counted": {
            "title": "Counted",
            "type": "boolean"
          },
          "delta_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta S"
          },
          "detected_at": {
            "format": "date-time",
            "title": "Detected At",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "pga_g": {
            "title": "Pga G",
            "type": "number"
          },
          "sensor_id": {
            "format": "uuid",
            "title": "Sensor Id",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "sensor_id",
          "detected_at",
          "pga_g",
          "delta_s",
          "counted"
        ],
        "title": "QuorumVoteOut",
        "type": "object"
      },
...
      "SeismicEventOut": {
        "description": "Fila de ``seismic_events`` (dato de red; epicentro aplanado a lon/lat).",
        "properties": {
          "depth_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Depth Km"
          },
          "detected_at": {
            "format": "date-time",
            "title": "Detected At",
            "type": "string"
          },
          "epicenter_lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Epicenter Lat"
          },
          "epicenter_lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Epicenter Lon"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "magnitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Magnitude"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "source",
          "magnitude",
          "epicenter_lon",
          "epicenter_lat",
          "depth_km",
          "detected_at",
          "meta"
        ],
        "title": "SeismicEventOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "TAKAB API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/events": {
      "get": {
        "description": "Lista eventos sísmicos (detected_at desc) con paginación keyset estable.",
        "operationId": "list_events_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Events"
      }
    },
    "/events/{event_id}": {
      "get": {
        "description": "Detalle del evento + sus votos de quórum. 404 si el evento no existe.",
        "operationId": "get_event_events__event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Event"
      }
    },
    "/evidence/{evidence_id}/download": {
      "post": {
        "description": "Presigned GET 300 s de un objeto de evidencia + huella en ``audit_log``.",
        "operationId": "download_evidence_evidence__evidence_id__download_post",
        "parameters": [
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Evidence Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedDownload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Evidence"
      }
    },
...
    "/incidents/{incident_id}/dictamens": {
      "get": {
        "description": "Cadena de dictámenes del incidente (más reciente primero). 404 si no visible.",
        "operationId": "list_dictamens_incidents__incident_id__dictamens_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictamenList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Dictamens"
      },
      "post": {
        "description": "Firma un dictamen: inserta fila nueva superseeding la última de la cadena.",
        "operationId": "sign_dictamen_incidents__incident_id__dictamens_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictamenSignIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictamenOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sign Dictamen"
      }
    },
    "/incidents/{incident_id}/evidence": {
      "get": {
        "description": "Lista la evidencia de un incidente visible para el tenant (RLS).",
        "operationId": "list_evidence_incidents__incident_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Evidence"
      }
    },
...
  }
}